+++ package/src/cjs/index.cjs
@@ -1,13 +1,13 @@
 /*
 * src/cjs-tools/index.cjs
 */
-const clearFirestoreData = require('./clearFirestoreData.cjs');
-const prime = require('./prime.cjs');
+//const clearFirestoreData = require('./clearFirestoreData.cjs');
+//const prime = require('./prime.cjs');
 
 const jestResolver = require('./jestResolver.cjs');
 
 module.exports = {
-  clearFirestoreData,
-  prime,
+  //clearFirestoreData,
+  //prime,
   jestResolver
 }