+++ package/src/config.js
@@ -3,16 +3,16 @@
 *
 * Provides access to emulator configuration
 */
-import fs from "fs"
+import { readFileSync } from 'fs'
 
 const fn = process.env["FIREBASE_JSON"] || './firebase.json';
 
 const firebaseJson = JSON.parse(
-  fs.readFileSync(fn, 'utf8')
+  readFileSync(fn, 'utf8')
 );
 
 const FIRESTORE_HOST = process.env["FIRESTORE_EMULATOR_HOST"] || (() => {
-  const port = firebaseJson.emulators.firestore.port;   // "6767"
+  const port = firebaseJson?.emulators?.firestore?.port;   // "6767"
   if (!port) {
     throw new Error(`Unable to get Firestore emulator port from '${fn}'`);
   }
@@ -21,7 +21,7 @@ const FIRESTORE_HOST = process.env["FIRESTORE_EMULATOR_HOST"] || (() => {
 })();
 
 const FUNCTIONS_URL = (() => {
-  let port = firebaseJson.emulators.functions?.port;   // "5002"
+  let port = firebaseJson?.emulators?.functions?.port;   // "5002"
   if (!port) {
     port = 5001;
     console.warning(`No 'emulators.functions.port' in '${fn}': using default (${port}).`);
@@ -29,7 +29,6 @@ const FUNCTIONS_URL = (() => {
 
   return `http://localhost:${port}`;
 })();
-
 
 export {
   FIRESTORE_HOST,