+++ package/src/firestore/index.js
@@ -3,8 +3,10 @@
 */
 import { dbUnlimited } from "./dbUnlimited.js"
 import { prime } from "./prime.js"
+import { clearAll } from "./clearAll.js"
 
 export {
   dbUnlimited,
-  prime
+  prime,
+  clearAll
 }