+++ package/src/projectId.js
@@ -1,8 +1,7 @@
 /*
-* src/projectId.js
+* src/firestore/projectId.js
 *
-* Dig the Firebase projectId (needed for showing Firestore data in the Emulator UI but also for initializing the
-* cloud Functions app - or is it?).  <<-- tbd. advice from Firebase requested, to better understand projects and apps
+* Dig the Firebase projectId (needed for accessing Firestore data).
 *
 * Our means:
 *   - 'GCLOUD_PROJECT' env.var. set by 'firebase emulators:exec', automatically