+++ sample/test-rules/docs.js
@@ -1,9 +1,9 @@
 /*
-* sample/test-rules/docs.cjs
+* sample/test-rules/docs.js
 *
 * Constant data used for the rules test cases
 *
-* Note: Needs to be '.cjs' because loaded from Jest 'globalSetup' (no ES modules support, yet).
+* REMOVE Note: Needs to be '.cjs' because loaded from Jest 'globalSetup' (no ES modules support, yet).
 */
 const anyPastDate = new Date();
 
@@ -86,5 +86,5 @@ const docs = {
   }
 };
 
-//export { docs }
-module.exports = docs;
+export { docs }
+//module.exports = docs;