+++ sample/test-rules/jest.config.js
@@ -4,5 +4,5 @@ import opts from '../jest.config.default.js'
 
 export default { ...opts,
   // Load docs, once at the beginning of the tests.
-  globalSetup: "./setup.jest.cjs"
+  globalSetup: "./setup.jest.js"
 }